remove map_traits header from containers repo from the HELICS public headers.  Do this by adding a second header federate_id_extra which has the links to map_traits but is not needed outside of the internal HELICS code
