+++ src/helics/application_api/ValueFederateManager.hpp
@@ -8,6 +8,7 @@ SPDX-License-Identifier: BSD-3-Clause
 
 #include "../common/GuardedTypes.hpp"
 #include "../core/Core.hpp"
+#include "../core/federate_id_extra.hpp"
 #include "Inputs.hpp"
 #include "Publications.hpp"
 #include "containers/DualMappedVector.hpp"
