+++ src/helics/core/BrokerBase.hpp
@@ -13,7 +13,7 @@ and some common methods used cores and brokers
 
 #include "ActionMessage.hpp"
 #include "containers/BlockingPriorityQueue.hpp"
-#include "federate_id.hpp"
+#include "federate_id_extra.hpp"
 #include <atomic>
 #include <memory>
 #include <string>
