+++ src/helics/core/FederateState.hpp
@@ -14,7 +14,6 @@ SPDX-License-Identifier: BSD-3-Clause
 #include "InterfaceInfo.hpp"
 #include "TimeDependencies.hpp"
 #include "containers/BlockingQueue.hpp"
-//#include "containers/DualMappedPointerVector.hpp"
 #include "core-data.hpp"
 #include "core-types.hpp"
 #include "helics-time.hpp"
@@ -105,7 +104,7 @@ class FederateState
     std::atomic_flag processing = ATOMIC_FLAG_INIT;  //!< the federate is processing
   private:
     /** a logging function for logging or printing messages*/
-    std::function<void (int, const std::string &, const std::string &)> loggerFunction;
+    std::function<void(int, const std::string &, const std::string &)> loggerFunction;
     std::function<std::string (const std::string &)> queryCallback;  //!< a callback for additional queries
     /** find the next Value Event*/
     Time nextValueTime () const;
@@ -291,7 +290,7 @@ class FederateState
     @details function must have signature void(int level, const std::string &sourceName, const std::string
     &message)
     */
-    void setLogger (std::function<void (int, const std::string &, const std::string &)> logFunction)
+    void setLogger (std::function<void(int, const std::string &, const std::string &)> logFunction)
     {
         loggerFunction = std::move (logFunction);
     }
